samples: bluetooth: peripheral_hids_keyboard: enable nfct for nRF54H20
Enabled the nfct node for the nRF54H20 DK in the DTS configuration of
the Peripheral HIDS Keyboard sample. The DTS node configuration for the
nfct node also requires configuration of the memory region attribute
so that the NFCT peripheral could be assigned to the specific core.

Ref: NCSDK-31691

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>
(cherry picked from commit 6a13c27834ccce83ed968828c42ca4072c27f902)